integrating salesmodel, service and controller
