--- back-end/controllers/checkoutController.js
@@ -1,9 +1,0 @@
-const { finishSale } = require('../services/checkoutService');
-
-const finishSalesController = async (req, res) => {
-  const sales = await finishSale(req.user, req.body);
-  console.log(sales);
-  return res.status(200).json(sales);
-};
-
-module.exports = { finishSalesController };
