+++ back-end/routers/index.js
@@ -3,6 +3,7 @@ const register = require('./register');
 const profile = require('./profile');
 const products = require('./products');
 const admin = require('./admin');
+const orders = require('./orders');
 const checkout = require('./checkout');
 
 module.exports = {
@@ -12,4 +13,5 @@ module.exports = {
   products,
   admin,
   checkout,
+  orders,
 };
